requirements.txt
websockets==12.0
pyaudio==0.2.14
python-dotenv==1.0.0

# 本地 ASR 依赖
torch>=2.0.0
openai-whisper>=20230918
numpy>=1.24.0

